refactor(PT-4175): rename step-props model, tighten onNext contract + docs, loading/error UX polish
